move Collision, Rendering and some Constants to a file
