+++ ArakaDev/ArakaDev.vcxproj
@@ -12,11 +12,16 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\Bullet.cpp" />
+    <ClCompile Include="..\Collision.cpp" />
     <ClCompile Include="..\Jilltest.cpp" />
+    <ClCompile Include="..\Rendering.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\Bullet.h" />
+    <ClInclude Include="..\Collision.h" />
+    <ClInclude Include="..\Constants.h" />
     <ClInclude Include="..\Enums.h" />
+    <ClInclude Include="..\Rendering.h" />
     <ClInclude Include="..\Structs.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
